feat: extract @gridplus/types package from SDK
Extract shared types and protocol constants into a new workspace package
for better code organization and reusability.

New package @gridplus/types includes:
- Protocol enums (LatticeResponseCode, LatticeMsgType, etc.)
- Protocol constants (ProtocolConstants)
- SDK constants (CURRENCIES)
- Shared types (KeyPair, KVRecords, WalletPath, etc.)
- Client types (Wallet, ActiveWallets, SignData, etc.)
- Firmware types (FirmwareConstants, FirmwareVersion)
- Message types (LatticeMessageHeader, LatticeSecureRequest, etc.)
- Sign types (TransactionRequest, SigningPayload, etc.)

SDK now re-exports from @gridplus/types, maintaining API compatibility.
SDK-specific types that reference the Client class remain in the SDK.
