+++ packages/sdk/package.json
@@ -55,6 +55,7 @@
   },
   "dependencies": {
     "@gridplus/btc": "workspace:*",
+    "@gridplus/types": "workspace:*",
     "@ethereumjs/common": "^10.0.0",
     "@ethereumjs/rlp": "^10.0.0",
     "@ethereumjs/tx": "^10.0.0",
