+++ packages/sdk/src/types/shared.ts
@@ -1,19 +1,10 @@
-import type { ec } from 'elliptic';
-
-export interface KVRecords {
-  [key: string]: string;
-}
-
-export interface EncrypterParams {
-  payload: Buffer;
-  sharedSecret: Buffer;
-}
-
-export type KeyPair = ec.KeyPair;
-
-export type WalletPath = [number, number, number, number, number];
-
-export interface DecryptedResponse {
-  decryptedData: Buffer;
-  newEphemeralPub: KeyPair;
-}
+/**
+ * Re-export shared types from @gridplus/types
+ */
+export type {
+  KVRecords,
+  EncrypterParams,
+  KeyPair,
+  WalletPath,
+  DecryptedResponse,
+} from '@gridplus/types';
