+++ packages/types/package.json
@@ -1,0 +1,46 @@
+{
+  "name": "@gridplus/types",
+  "version": "0.1.0",
+  "type": "module",
+  "description": "Shared types for GridPlus SDK",
+  "scripts": {
+    "build": "tsup",
+    "lint": "biome check src",
+    "lint:fix": "biome check --write src",
+    "typecheck": "tsc --noEmit"
+  },
+  "files": [
+    "dist"
+  ],
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.mjs",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
+    },
+    "./package.json": "./package.json"
+  },
+  "types": "./dist/index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/GridPlus/gridplus-sdk.git",
+    "directory": "packages/types"
+  },
+  "dependencies": {
+    "elliptic": "6.6.1",
+    "viem": "^2.37.8"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "^1.9.0",
+    "@types/elliptic": "^6.4.18",
+    "@types/node": "^24.10.4",
+    "tsup": "^8.5.0",
+    "typescript": "^5.9.2"
+  },
+  "license": "MIT",
+  "engines": {
+    "node": ">=20"
+  }
+}
