+++ packages/types/src/index.ts
@@ -1,0 +1,23 @@
+// Protocol enums and constants
+export * from './protocol';
+
+// SDK constants
+export * from './constants';
+
+// Shared types
+export * from './shared';
+
+// Message types
+export * from './messages';
+
+// Secure message types
+export * from './secureMessages';
+
+// Client-related types
+export * from './client';
+
+// Firmware types
+export * from './firmware';
+
+// Signing types
+export * from './sign';
