+++ packages/types/src/shared.ts
@@ -1,0 +1,19 @@
+import type { ec } from 'elliptic';
+
+export interface KVRecords {
+  [key: string]: string;
+}
+
+export interface EncrypterParams {
+  payload: Buffer;
+  sharedSecret: Buffer;
+}
+
+export type KeyPair = ec.KeyPair;
+
+export type WalletPath = [number, number, number, number, number];
+
+export interface DecryptedResponse {
+  decryptedData: Buffer;
+  newEphemeralPub: KeyPair;
+}
